requirements.txt
EasyTkinter==1.1.0
Pillow==8.0.1
opencv-python==4.4.0.46
serial==0.0.97
RPi.GPIO==0.7.0
gpiozero==1.5.1